Revert "[BZ 845277] - quick fix: removed the doctype declaration and removed overcasting to int"
This reverts commit 8000e712dd98fd8ee8e1ff52940e15b07f11e782.

+++ modules/enterprise/gui/coregui/src/main/java/org/rhq/enterprise/gui/coregui/client/inventory/common/AbstractMetricGraphView.java
@@ -312,7 +312,7 @@ public abstract class AbstractMetricGraphView extends LocatableVLayout {
 
         for (MeasurementDataNumericHighLowComposite d : data) {
             if (!Double.isNaN(d.getValue())) {
-                handler.add(new DataPoint(d.getTimestamp(), d.getValue()));
+                handler.add(new DataPoint(d.getTimestamp(), (int) d.getValue()));
             }
         }
